Closes #23 and more updates
Removed .registration = TRUE for C calls, because with the new version
of roxygen2 it was not reloading the package correctly. Furthermore, now
it is no longer necessary (I am not sure about what has happened and why
earlier this was necessary. We will probably learn more about this when
doing #19).